Enviando alterações de exercicio com retorno msg no delete

+++ src/main/java/br/com/gymfy/services/UsuarioService.java
@@ -4,8 +4,11 @@ package br.com.gymfy.services;
 import br.com.gymfy.entities.Exercicio;
 import br.com.gymfy.entities.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import br.com.gymfy.repositories.UsuarioRepository;
+import org.springframework.web.server.ResponseStatusException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -47,7 +50,12 @@ public class UsuarioService {
     }
 
     public void deletar(Integer id) {
-        usuarioRepository.deleteById(id);
+        Optional<Usuario> usuario = usuarioRepository.findById(id);
+        if (usuario.isPresent()) {
+            usuarioRepository.deleteById(id);
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Usuário com ID " + id + " não encontrado.");
+        }
     }
 
     public Usuario update (Integer id, Usuario usuario){
